+++ stroom-core-client/src/main/java/stroom/importexport/client/presenter/DependenciesUiHandlers.java
@@ -1,0 +1,7 @@
+package stroom.importexport.client.presenter;
+
+import com.gwtplatform.mvp.client.UiHandlers;
+
+public interface DependenciesUiHandlers extends UiHandlers {
+    void changeQuickFilter(String name);
+}